requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
httpx
python-dotenv
email-validator
motor>=3.7.1,<4.0.0
python-jose
pymongo>=4.13.2,<5.0.0
passlib[bcrypt]
python-multipart
openai
tenacity>=8.2.3

# RAG Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
sentence-transformers>=2.2.2
pypdf>=3.0.0

# MongoDB
pymongo[srv]>=4.5.0
pymongo[encryption]>=4.5.0